Updated cart

+++ src/main/java/com/example/hop_oasis/dto/CartDto.java
@@ -2,6 +2,7 @@ package com.example.hop_oasis.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @AllArgsConstructor
@@ -11,5 +12,5 @@ import java.util.List;
 @Setter
 public final class CartDto {
     private List<CartItemDto> items;
-    private double priceForAll;
+    private BigDecimal priceForAll;
 }
